pollute the activate method less
