+++ src/extension_utils/utils.ts
@@ -4,6 +4,7 @@
 import { ExtensionContext, MessageItem, Uri, window } from "vscode";
 import * as path from "path";
 import { CONSTANTS, DialogResponses, USER_CODE_NAMES } from "../constants";
+import { DependencyChecker } from "./dependencyChecker";
 
 // tslint:disable-next-line: export-name
 export const getPathToScript = (
@@ -35,4 +36,32 @@ export const showPrivacyModal = (okAction: () => void) => {
         okAction();
       }
     })
+}
+
+export const checkPythonDependency = async () => {
+  const dependencyChecker: DependencyChecker = new DependencyChecker();
+  const result = await dependencyChecker.checkDependency(CONSTANTS.DEPENDENCY_CHECKER.PYTHON);
+  return result.payload;
+}
+
+export const setPythonExectuableName = async () => {
+  // Find our what command is the PATH for python
+  let executableName: string = "";
+  const dependencyCheck = await checkPythonDependency();
+  if (dependencyCheck.installed) {
+    executableName = dependencyCheck.dependency;
+  } else {
+    window.showErrorMessage(CONSTANTS.ERROR.NO_PYTHON_PATH,
+      DialogResponses.INSTALL_PYTHON)
+      .then((selection: MessageItem | undefined) => {
+        if (selection === DialogResponses.INSTALL_PYTHON) {
+          const okAction = () => {
+            open(CONSTANTS.LINKS.DOWNLOAD_PYTHON);
+          };
+          showPrivacyModal(okAction);
+        }
+      });
+  }
+
+  return executableName;
 }
